Change the dir structure
