+++ src/java/CMakeLists.txt
@@ -78,7 +78,7 @@ if(TRITON_ENABLE_JAVA_HTTP)
   include(UseJava)
   set(PROJECT_TARGET_DIR "${CMAKE_CURRENT_SOURCE_DIR}/target")
   set(PROJECT_JAR "java-sdk.jar")
-  set(EXAMPLES_PATH "src/main/java/com/nvidia/triton/contrib")
+  set(EXAMPLES_PATH "src/main/java/triton/client")
   add_custom_command(
     OUTPUT "${PROJECT_TARGET_DIR}/${PROJECT_JAR}"
     COMMAND mkdir -p ${CMAKE_INSTALL_PREFIX}/java/${EXAMPLES_PATH}
