+++ src/java/src/main/java/triton/client/InferenceException.java
@@ -24,44 +24,23 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-package com.nvidia.triton.contrib.pojo;
+package triton.client;
+
+import triton.client.pojo.ResponseError;
 
 /**
- * Data types from
+ * Universal exceptions of triton client.
  */
-public enum DataType {
+public class InferenceException extends Exception {
+    public InferenceException(ResponseError err) {
+        super(err.getError());
+    }
 
-    BOOL(1, false),
-    INT8(1, true),
-    INT16(2, true),
-    INT32(4, true),
-    INT64(8, true),
-    UINT8(1, false),
-    UINT16(2, false),
-    UINT32(4, false),
-    UINT64(8, false),
-    FP16(2, false),
-    FP32(4, false),
-    FP64(8, false),
-    BYTES(-1, false);
+    public InferenceException(String message) {
+        super(message);
+    }
 
-    public final int numByte;
-    public final boolean singed;
-
-    DataType(int numByte, boolean singed) {
-        this.numByte = numByte;
-        this.singed = singed;
+    public InferenceException(Throwable cause) {
+        super(cause);
     }
 }
-
-
-
-
-
-
-
-
-
-
-
-
