+++ src/java/src/main/java/triton/client/examples/TritonPerf.java
@@ -24,18 +24,18 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-package com.nvidia.triton.contrib.examples;
+package triton.client.examples;
 
 import java.util.Arrays;
 import java.util.List;
 
-import com.nvidia.triton.contrib.InferInput;
-import com.nvidia.triton.contrib.InferRequestedOutput;
-import com.nvidia.triton.contrib.InferResult;
-import com.nvidia.triton.contrib.InferenceServerClient;
-import com.nvidia.triton.contrib.InferenceServerClient.InferArguments;
-import com.nvidia.triton.contrib.endpoint.FixedEndpoint;
-import com.nvidia.triton.contrib.pojo.DataType;
+import triton.client.InferInput;
+import triton.client.InferRequestedOutput;
+import triton.client.InferResult;
+import triton.client.InferenceServerClient;
+import triton.client.InferenceServerClient.InferArguments;
+import triton.client.endpoint.FixedEndpoint;
+import triton.client.pojo.DataType;
 import com.google.common.collect.Lists;
 import com.google.common.util.concurrent.AtomicDouble;
