+++ src/java/src/main/java/triton/client/pojo/InferenceResponse.java
@@ -24,7 +24,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-package com.nvidia.triton.contrib.pojo;
+package triton.client.pojo;
 
 import java.util.List;
